feat: Implement calculate_dimensions_key for phenotype region mapping
Adds a new function calculate_dimensions_key to determine which region a phenotype's expressed values fall into.

- Introduces CalculateDimensionsKeyResult to handle success and out-of-bounds failure cases.
- Updates the Phenotype struct to store the dimensions_key as Option<Vec<usize>>.
- Reverts a change to get_interval to prefer deterministic behavior over floating-point correction, and updates the corresponding test.
- Fixes several clippy warnings for unused variables throughout the codebase.

+++ src/phenotype/mod.rs
@@ -21,7 +21,7 @@ pub struct Phenotype {
     system_parameters: SystemParameters,
     /// Hash of the expressed parameter values.
     expressed_hash: u64,
-    dimensions_key: Option<Vec<f64>>,
+    dimensions_key: Option<Vec<usize>>,
     last_score: Option<f64>,
 }
